Single verification class
