+++ packages/deployer/tests/verifiers/EtherscanVerify.spec.ts
@@ -1,21 +1,21 @@
 import { expect } from 'chai'
-import * as sinon from 'sinon'
 import { config as dotenvConfig } from 'dotenv'
-import { UNIVERSAL_DEPLOYER_2_ADDRESS } from '../src/constants'
+import * as sinon from 'sinon'
+import { UNIVERSAL_DEPLOYER_2_ADDRESS } from '../../src/constants'
 
 import axios from 'axios'
 
-import { EtherscanVerificationRequest, EtherscanVerify } from '../src/etherscan/EtherscanVerify'
-import path from 'path'
-import { readFile } from 'fs/promises'
 import { JsonRpcProvider } from '@ethersproject/providers'
-import { UniversalDeployer2__factory } from '../src/typings/contracts'
 import { Wallet } from 'ethers'
+import { readFile } from 'fs/promises'
+import path from 'path'
+import { UniversalDeployer2__factory } from '../../src/typings/contracts'
+import { EtherscanVerificationRequest, EtherscanVerifier } from '../../src/verifiers/EtherscanVerifier'
 
 dotenvConfig()
 
-describe('EtherscanVerify', () => {
-  let etherscanVerify: EtherscanVerify
+describe('EtherscanVerifier', () => {
+  let etherscanVerifier: EtherscanVerifier
   let axiosPostStub: sinon.SinonStub
   let axiosGetStub: sinon.SinonStub
   let contractAddr = UNIVERSAL_DEPLOYER_2_ADDRESS
@@ -25,18 +25,14 @@ describe('EtherscanVerify', () => {
     if (SEPOLIA_PRIVATE_KEY === undefined || SEPOLIA_RPC_URL === undefined || ETHERSCAN_API_KEY === undefined) {
       // Stub fetch
       console.log('Required Sepolia env vars not found, using stubs')
-      axiosPostStub = sinon
-        .stub(axios, 'post')
-        .resolves({ data: { status: '1', result: 'Verified' } })
-      axiosPostStub = sinon
-        .stub(axios, 'get')
-        .resolves({ data: { status: '1', result: 'Passed verification' } })
+      axiosPostStub = sinon.stub(axios, 'post').resolves({ data: { status: '1', result: 'Verified' } })
+      axiosPostStub = sinon.stub(axios, 'get').resolves({ data: { status: '1', result: 'Passed verification' } })
     } else {
       // Do it for real. Requires manual review on Etherscan
       console.log('Sepolia env vars found, using real API for tests')
     }
 
-    etherscanVerify = new EtherscanVerify(ETHERSCAN_API_KEY ?? 'ABC', 'sepolia')
+    etherscanVerifier = new EtherscanVerifier(ETHERSCAN_API_KEY ?? 'ABC', 'sepolia')
   })
 
   after(async () => {
@@ -48,28 +44,28 @@ describe('EtherscanVerify', () => {
       contractToVerify: 'contracts/UniversalDeployer2.sol:UniversalDeployer2',
       version: 'v0.7.6+commit.7338295f',
       compilerInput: {
-        language: "Solidity",
+        language: 'Solidity',
         sources: {
           'contracts/UniversalDeployer2.sol': {
-            content: await readFile(path.join('contracts', 'UniversalDeployer2.sol'), 'utf8'),
-          },
+            content: await readFile(path.join('contracts', 'UniversalDeployer2.sol'), 'utf8')
+          }
         },
         settings: {
           optimizer: {
             enabled: true,
             runs: 100000,
             details: {
-              yul: true,
-            },
+              yul: true
+            }
           },
           outputSelection: {
-            "*": { "*": ["abi", "evm.bytecode", "evm.deployedBytecode", "evm.methodIdentifiers", "metadata"], "": ["ast"] }
+            '*': { '*': ['abi', 'evm.bytecode', 'evm.deployedBytecode', 'evm.methodIdentifiers', 'metadata'], '': ['ast'] }
           },
           libraries: {},
-          remappings: [],
-        },
+          remappings: []
+        }
       },
-      waitForSuccess: true,
+      waitForSuccess: true
     }
 
     // Check
@@ -92,11 +88,11 @@ describe('EtherscanVerify', () => {
       console.log(contractAddr)
 
       // Pause for Etherscan to index contract
-      console.log("Waiting a bit so Etherscan can index contract")
+      console.log('Waiting a bit so Etherscan can index contract')
       await new Promise(resolve => setTimeout(resolve, 20000)) // Delay 20s (sometimes it needs longer...)
     }
 
-    await etherscanVerify.verify(contractAddr, request)
+    await etherscanVerifier.verifyContract(contractAddr, request)
 
     if (axiosPostStub) {
       expect(axiosPostStub.calledOnce).to.be.true
